Refactor of loading logic.
The logic to dynamically load the actual image data reworked to be
acutally usable.

For that the structure of the Thumbnail struct was changed. It now
contains an enum which either holds a File handle or the actual image
data in the form of a image::DynamicImage instance.

This required some logic changes on the Operation::apply functions
which then where combined with a signature change.
Operation::apply now directly takes a reference to a DynamicImage
instead of taking the whole Thumbnail instance.
This allowed the code to be much slimer and removes a lot if ownership
problems.