Refactored Thumbnail struct
Moved all actual image data into another sub struct.
This is done to separate the ops vector field from the image data.
This allows us to have only a vector of ThumbnailData without the ops
field in ThumbnailCollection.

The reason for this being, that a vector of a trait cannot be Sized and
therefore not be used in multi threading. Also having n instances of an
emtpy unneeded vector is also unnecessary.