BUGFIX: Image reader wasn't given the image format.
When the Reader from the image create is manually created, it does not
guess the image format. It wasn't really supposed to anyway since that
has already been done earlier, and the information was available, but
just not passed on to the Reader instance.

This caused the decode call to fail, since the reader could not
determine an image data type.