BUGFIX: Target destination path ignored.
The store method used std::path::Path::is_file() to check whether the
given path was a path to a file. However this checks whether the file
actually exists, which was not the desired behaviour.

In fact it is not even exactly possible to determine from a path alone
whether it points to a file or a directory.

The function has now been extended with a set of rules to determine what
to do depending on how the target path looks like.